requirements.txt
streamlit==1.29.0
google-generativeai==0.8.3
python-dotenv==1.0.0
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0
httpx==0.28.1
reportlab==4.0.9